fix(AxisCreator): add remove axis method and custom axis1 and axis2
The Axis Creator now has a custom axis1 and axis2 which is used by the
prefabs instead of relying on the Horizontal and Vertical axes provided
by Unity as these are not clean axis settings.

The AxisCreator editor window also allows deleting of existing Tilia
axis inputs so its easier to upgrade.

The incorrect negative multiplier in the controller mappings has also
been removed now as the inversion happens on the axis setting in the
Unity Input Manager.