fix(AxisCreator): prevent null exception if no axis found
There was an issue when no axis was found it would throw a null
exception as the `Next` call would try and iterate through to a
null element. This is resolved by just using the resulting bool
from the `Next` call to ensure whether it can proceed wit the
additional check.